pyproject.toml
[tool.poetry]
name = "bite-the-jaw"
version = "0.1.0"
description = ""
authors = ["Ava Yazdany <avayaz1337@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
PyJWT = "^2.3.0"
django-cors-headers = "^3.11.0"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.950"
autopep8 = "^1.6.0"
pylint = "^2.13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
